+++ src/com/arcsoft/refocus/refocus/VideoRefocus.java
@@ -7,7 +7,7 @@ public class VideoRefocus {
 		System.loadLibrary("arcsoft_dualcam_refocus_wrap");
 	}
 
-	private int mEngineHandler = 0;
+	private long mEngineHandler = 0;
 
 	private boolean IsInited() {
 		return (mEngineHandler != 0);
@@ -69,17 +69,17 @@ public class VideoRefocus {
 	/***********************************************************************/
 	/********************** HDR Native JNI Define **********************/
 	/***********************************************************************/
-	private native int VideoRefocus_Init();
+	private native long VideoRefocus_Init();
 
-	private native void VideoRefocus_Uninit(int hEngine);
+	private native void VideoRefocus_Uninit(long hEngine);
 	
-	private native int VideoRefocus_SetCameraImageInfo(int hEngine, int leftW, int leftH, int rightW, int rightH);
+	private native int VideoRefocus_SetCameraImageInfo(long hEngine, int leftW, int leftH, int rightW, int rightH);
 	
-	private native int VideoRefocus_SetImageDegree(int hEngine, int degree);
+	private native int VideoRefocus_SetImageDegree(long hEngine, int degree);
 	
-	private native int VideoRefocus_SetCalibrationData(int hEngine, byte[] caliData, int length);
+	private native int VideoRefocus_SetCalibrationData(long hEngine, byte[] caliData, int length);
 	
-	private native int VideoRefocus_SetParam(int hEngine, int pointX, int pointY, int blurIntensity, int bOn);
+	private native int VideoRefocus_SetParam(long hEngine, int pointX, int pointY, int blurIntensity, int bOn);
 	
-	private native byte[] VideoRefocus_Process(int hEngine, byte[] leftData, int leftlength, byte[] rightData, int rightlength, int mainWidth, int mainHeight, int auxWidth, int auxHeight);
+	private native byte[] VideoRefocus_Process(long hEngine, byte[] leftData, int leftlength, byte[] rightData, int rightlength, int mainWidth, int mainHeight, int auxWidth, int auxHeight);
 }
